backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
openai
langgraph==0.2.43
langchain==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.19
python-dotenv==1.0.1
pydantic==2.9.0
